test now has its own Project.toml
