+++ Project.toml
@@ -52,9 +52,3 @@ SpecialFunctions = "2"
 SplittableRandoms = "0.1"
 StatsBase = "0.33"
 julia = "1.6"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
